Resolved warnings. Moved color-scheme store to stores file.
