+++ static/app.css
@@ -35,7 +35,7 @@
 	--secondary-color: #d22d19;
 	--main-color: #fff;
 	--content-color: #373a3d;
-	--tertiary-color: #DE9B8A;
+	--tertiary-color: #de9b8a;
 	--tertiary-font-color: #000;
 	--title-color: #2b2b2b;
 }
